FIX: reorganize imports and add GenerateAndSetToken function for JWT token generation
