+++ middlewares/authMiddleware.go
@@ -2,11 +2,12 @@ package middlewares
 
 import (
 	"fmt"
-	"github.com/dgrijalva/jwt-go"
-	"github.com/gin-gonic/gin"
 	"log"
 	"net/http"
 	"os"
+
+	"github.com/dgrijalva/jwt-go"
+	"github.com/gin-gonic/gin"
 )
 
 // AuthMiddleware validates JWT tokens and protects routes
