+++ routes/authRoutes.go
@@ -3,8 +3,10 @@ package routes
 import (
 	"civicsync-be/controllers"
 	"civicsync-be/middlewares"
+
 	"github.com/gin-gonic/gin"
 )
+
 // AuthRoutes sets up the authentication routes
 func AuthRoutes(r *gin.Engine) {
 	auth := r.Group("/api/auth")
